Condition added

+++ services/lesson.js
@@ -1,5 +1,6 @@
 const Lesson = require('~/models/lesson')
 const uploadService = require('~/services/upload')
+const { createForbiddenError } = require('~/utils/errorsHelper')
 const { ATTACHMENT } = require('~/consts/upload')
 
 const lessonService = {
@@ -13,7 +14,17 @@ const lessonService = {
     return await Lesson.create({ author, title, description, attachments: fileUrls })
   },
 
-  deleteLesson: async (id) => {
+  deleteLesson: async (id, currentUser) => {
+    const { id: currentUserId } = currentUser
+
+    const lesson = await Lesson.findById(id)
+
+    const lessonAuthor = lesson.author.toString()
+
+    if (lessonAuthor !== currentUserId) {
+      throw createForbiddenError()
+    }
+
     await Lesson.findByIdAndRemove(id).exec()
   }
 }
